Update Ski-ScaleResolutions.plugin.js
Hopefully the final clean up before I work on adding a button to the GUI.

Removed Wide and Tall options, now you just have to use numbers lower than 1, or math to get tall images.

EX: For 2:3 Images, that are Tall
USE: .666 or 2/3 or 512/768

They all equal the same thing.