Update Ski-ScaleResolutions.plugin.js
Another change to make it easier to get different max resolutions in the all_resolutions function. and it also will round numbers to the closest 64.

Also added option to have the scaled_resolutions to either override or append to the resolution list. In case you want to make the list more manage able.